backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
httpx==0.25.1
pydantic==2.4.2
python-dotenv==1.0.0
websockets==12.0
aiosqlite==0.19.0
web3==7.13.0
requests==2.32.5
